Endring av kodeeksempler for CopyButton
