--- aksel.nav.no/website/pages/eksempler/copybutton/action.tsx
@@ -1,18 +1,0 @@
-import { CopyButton } from "@navikt/ds-react";
-import { withDsExample } from "@/web/examples/withDsExample";
-
-const Example = () => {
-  return <CopyButton copyText="3.14" variant="action" />;
-};
-
-// EXAMPLES DO NOT INCLUDE CONTENT BELOW THIS LINE
-export default withDsExample(Example);
-
-/* Storybook story */
-export const Demo = {
-  render: Example,
-};
-
-export const args = {
-  index: 1,
-};
