--- aksel.nav.no/website/pages/eksempler/copybutton/small.tsx
@@ -1,23 +1,0 @@
-import { CopyButton, HStack } from "@navikt/ds-react";
-import { withDsExample } from "@/web/examples/withDsExample";
-
-const Example = () => {
-  return (
-    <HStack gap="3">
-      <CopyButton size="small" copyText="3.14" />
-      <CopyButton size="small" copyText="3.14" text="Kopier" />
-    </HStack>
-  );
-};
-
-// EXAMPLES DO NOT INCLUDE CONTENT BELOW THIS LINE
-export default withDsExample(Example);
-
-/* Storybook story */
-export const Demo = {
-  render: Example,
-};
-
-export const args = {
-  index: 2,
-};
