+++ aksel.nav.no/website/pages/eksempler/copybutton/tooltip.tsx
@@ -20,5 +20,5 @@ export const Demo = {
 
 export const args = {
   index: 8,
-  desc: "Tooltip kan hjelpe med å gi mer kontekst om hva man kopierer. Vi har testet dette med skjermleser og funnet ut at statisk tekst som 'kopier ...' fungerer bedre enn dynamisk tekst ved klikk i tooltip.",
+  desc: "Tooltip kan hjelpe med å gi mer kontekst om hva man kopierer. Testing med skjermleser har vist at endringer av teksten i Tooltip ikke blir fanget opp av skjermlesere. Teksten bør derfor være statisk.",
 };
